convert squeak time to a reusable component
