+++ frontend/src/components/SqueakThreadItem/SqueakThreadItem.js
@@ -5,8 +5,6 @@ import {
   Grid,
   Box,
   Link,
-  Tooltip,
-  Button,
 } from '@material-ui/core';
 import { useHistory } from 'react-router-dom';
 
@@ -14,14 +12,10 @@ import LockIcon from '@material-ui/icons/Lock';
 import DownloadIcon from '@material-ui/icons/CloudDownload';
 
 // styles
-import moment from 'moment';
 import useStyles from './styles';
 
 import SqueakActionBar from '../SqueakActionBar';
-
-import {
-  getBlockDetailUrl,
-} from '../../bitcoin/blockexplorer';
+import SqueakTime from '../SqueakTime';
 
 import {
   goToSqueakPage,
@@ -40,9 +34,6 @@ export default function SqueakThreadItem({
 
   const history = useHistory();
 
-  const blockDetailUrl = () =>
-    // return "https://blockstream.info/testnet/block/" + squeak.getBlockHash();
-    getBlockDetailUrl(squeak.getBlockHash(), network);
   const onAddressClick = (event) => {
     event.preventDefault();
     event.stopPropagation();
@@ -103,37 +94,6 @@ export default function SqueakThreadItem({
           ? SqueakUnlockedContent()
           : SqueakLockedContent()}
       </>
-    );
-  }
-
-  function SqueakTime() {
-    if (!squeak) {
-      return (
-        <Box color="secondary.main" fontWeight="fontWeightBold">
-          Unknown time
-        </Box>
-      );
-    }
-    const squeakBlockTime = moment(squeak.getBlockTime()*1000)
-    return (
-      <Box color="secondary.main" fontWeight="fontWeightBold">
-        <Tooltip title={squeakBlockTime.toString()}>
-          <Button color="secondary" style={{fontWeight: "bold", textTransform: "lowercase"}}>{squeakBlockTime.fromNow()}</Button>
-        </Tooltip>
-        <span> </span>
-        (Block
-        <Link
-          href={blockDetailUrl()}
-          target="_blank"
-          rel="noopener"
-          onClick={(event) => event.stopPropagation()}
-        >
-          <span> </span>
-          #
-          {squeak.getBlockHeight()}
-        </Link>
-        )
-      </Box>
     );
   }
 
@@ -206,7 +166,11 @@ export default function SqueakThreadItem({
         alignItems="flex-start"
       >
         <Grid item>
-          {SqueakTime()}
+          <SqueakTime
+            hash={hash}
+            squeak={squeak}
+            network={network}
+          />
         </Grid>
       </Grid>
       {showActionBar
