+++ frontend/src/components/SqueakTime/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "SqueakTime",
+  "version": "0.0.0",
+  "private": true,
+  "main": "SqueakTime.js"
+}
